feat: add support for options handling and an action button
This PR adds support for specifying specific options when processing json input. If any items have options, the left/right buttons turn into option selectors for the whole list (this is a no-op for items that have no options). The item name and the currently selected option are delimited by a `:`. The default selected option is at index 0, and this can be changed via `selected_option` (invalid values will be corrected, with a message on stdout).

Any list item can have the optional field `enabled` set to `false` (default: `true`), which will disable the ability to change the option value. Items can have an optional field `supports_enabling` set to `true` (default: false) to provide the ability to enable or disable an item. Disabled items will be displayed in gray. A new `--enabled-button` (default: Y) flag _must_ be set in order to provide functionality for enabling and disabling items.

Additionally, a new flag named `--stdout-value` can be used to decide whether to output the currently `selected` item (default) or the entire `state`. If the entire state is used, the output is written as pretty-printed json. If the `--stdout-value` flag is set to `state`, then the state is written to stdout regardless of the exit code.

Finally, an action key is introduced. This can be mapped to any of A/B/X/Y via the `--action-button` flag (default: X), and custom text can be specified for the action (default: ACTION). The selected value will also be output to stdout if `--stdout-value` flag is set to `selected`.

Closes #7